requirements.txt
portia-sdk-python[all]
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-json-logger==2.0.7
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.29.0
redis==5.2.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
celery[redis]==5.4.0
flower==2.0.1
pyjwt==2.10.1
cryptography==44.0.0
requests==2.32.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
slack-sdk==3.34.0
notion-client==2.2.1
jira==3.8.0
stripe==12.2.0
sendgrid==6.12.1
twilio==9.7.1
supabase==2.8.1
httpx>=0.24,<0.28